\name{aggregate.SpatialPointsDataFrameListZoo}
\alias{aggregate.SpatialPointsDataFrameListZoo}
\title{aggregate.SpatialPointsDataFrameListZoo}
\usage{
  aggregate.SpatialPointsDataFrameListZoo(x, by, ...)
}
\arguments{
  \item{x}{TODO}

  \item{by}{TODO}

  \item{...}{TODO}
}
\description{
  aggregate.SpatialPointsDataFrameListZoo
}

