\name{as.SpatialPointsDataFrameList}
\alias{as.SpatialPointsDataFrameList}
\title{as...}
\usage{as.SpatialPointsDataFrameList(SpatialPointsDataFrame, list, by, all=FALSE)}
\description{as.SpatialPointsDataFrameList}
\arguments{\item{SpatialPointsDataFrame}{TODO}
\item{list}{TODO}
\item{by}{TODO}
\item{all}{TODO}}
