\name{as.SpatialPointsDataFrameListZoo}
\alias{as.SpatialPointsDataFrameListZoo}
\title{as...}
\usage{as.SpatialPointsDataFrameListZoo(SpatialPointsDataFrame, list, by=NA, all=FALSE)}
\description{as.SpatialPointsDataFrameListZoo}
\arguments{\item{SpatialPointsDataFrame}{TODO}
\item{list}{TODO}
\item{by}{TODO}
\item{all}{Logical.}}
