\name{as.data.frame.SpatialPointsDataFrameList}
\alias{as.data.frame.SpatialPointsDataFrameList}
\title{as.data.frame.SpatialPointsDataFrameList}
\usage{
  as.data.frame.SpatialPointsDataFrameList(x, ...)
}
\arguments{
  \item{x}{TODO}

  \item{...}{TODO}
}
\description{
  as.data.frame.SpatialPointsDataFrameList
}

