\name{as.Date.SpatialPointsDataFrameListZoo}
\alias{as.Date.SpatialPointsDataFrameListZoo}
\title{as.Date.SpatialPointsDataFrameListZoo}
\usage{
  as.Date.SpatialPointsDataFrameListZoo(x, ...)
}
\arguments{
  \item{x}{TODO}

  \item{...}{TODO}
}
\description{
  as.Date.SpatialPointsDataFrameListZoo
}

