\name{as.list.SpatialPointsDataFrameList}
\alias{as.list.SpatialPointsDataFrameList}
\title{as.list.SpatialPointsDataFrameList}
\usage{
  as.list.SpatialPointsDataFrameList(x, ...)
}
\arguments{
  \item{x}{TODO}

  \item{...}{TODO}
}
\description{
  as.list.SpatialPointsDataFrameList
}

