\name{as.SpatialPointsDataFrameList}
\alias{as.SpatialPointsDataFrameList}
\title{as.SpatialPointsDataFrameList}
\usage{
  as.SpatialPointsDataFrameList(SpatialPointsDataFrame,
    list, by, all = FALSE)
}
\arguments{
  \item{SpatialPointsDataFrame}{TODO}

  \item{list}{TODO}

  \item{by}{TODO}

  \item{all}{TODO}
}
\description{
  as.SpatialPointsDataFrameList
}

