\name{as.SpatialPointsDataFrameListZoo}
\alias{as.SpatialPointsDataFrameListZoo}
\title{as.SpatialPointsDataFrameListZoo}
\usage{
  as.SpatialPointsDataFrameListZoo(SpatialPointsDataFrame,
    list, by = NA, all = FALSE)
}
\arguments{
  \item{SpatialPointsDataFrame}{TODO}

  \item{list}{TODO}

  \item{by}{TODO}

  \item{all}{Logical.}
}
\description{
  as.SpatialPointsDataFrameListZoo
}

