\name{check_list_classes}
\alias{check_list_classes}
\title{check_list_classes}
\usage{
  check_list_classes(list, class_to_check,
    check_inheritance = FALSE)
}
\arguments{
  \item{list}{TODO}

  \item{class_to_check}{TODO}

  \item{check_inheritance}{TODO}
}
\description{
  check_list_classes
}

