\name{crop_extents_synced}
\alias{crop_extents_synced}
\title{crop_extents_synced...}
\usage{crop_extents_synced(x, y, borderx=0, bordery=0, force_window_odd=TRUE,
    verbose=TRUE, ...)}
\description{crop_extents_synced}
\arguments{\item{x}{TODO}
\item{y}{TODO}
\item{borderx}{x border in pixels}
\item{bordery}{y border in pixels}
\item{...}{TODO}}
