\name{extract_overlap_rasters}
\alias{extract_overlap_rasters}
\title{extract_overlap_rasters}
\usage{
  extract_overlap_rasters(raster1, raster2)
}
\arguments{
  \item{raster1}{TODO}

  \item{raster2}{TODO}
}
\description{
  Coerces two rasters to have a shared subset of their
  common overlap.
}
\examples{
TODO
}
\author{
  Jonathan A. Greenberg \email{STARStools@estarcion.net}
}

