\name{index.SpatialPointsDataFrameListZoo}
\alias{index.SpatialPointsDataFrameListZoo}
\title{index...}
\usage{index.SpatialPointsDataFrameListZoo(x, SpatialDataFrameList=TRUE, ...)}
\description{index.SpatialPointsDataFrameListZoo}
\arguments{\item{x}{TODO}
\item{SpatialDataFrameList}{TODO}
\item{...}{TODO}}
