\name{index.SpatialPointsDataFrameListZoo}
\alias{index.SpatialPointsDataFrameListZoo}
\title{index.SpatialPointsDataFrameListZoo}
\usage{
  index.SpatialPointsDataFrameListZoo(x,
    SpatialDataFrameList = TRUE, ...)
}
\arguments{
  \item{x}{TODO}

  \item{SpatialDataFrameList}{TODO}

  \item{...}{TODO}
}
\description{
  index.SpatialPointsDataFrameListZoo
}

