\name{kernel_maker}
\alias{kernel_maker}
\title{kernel_maker}
\usage{
  kernel_maker(kernelx = 3, kernely = 3, kernelz = 1,
    shape = "circular", outformat = "ENVI",
    outkernelfname = "mykernel.ker")
}
\arguments{
  \item{kernelx}{TODO}

  \item{kernely}{TODO}

  \item{kernelz}{TODO}

  \item{shape}{TODO}

  \item{outformat}{TODO}

  \item{outkernelfname}{TODO}
}
\description{
  kernel_maker
}

