\name{rstarspan}
\alias{rstarspan}
\title{Extracts raster data at vector locations.}
\usage{rstarspan(vectors, rasters, outformat="SpatialDataFrame")}
\description{Extracts raster data at vector locations.}
\arguments{\item{vectors}{a vector of superclass "Spatial" or a list of vectors.}
\item{rasters}{a raster of superclass "Raster" or a list of rasters.}
\item{outformat}{}}
