\name{tahoe_lidar_bareearth.tif}
\alias{tahoe_lidar_bareearth.tif}
\title{Lidar-derived bare earth digital elevation model from the Lake Tahoe Basin.}
\description{Lidar-derived bare earth digital elevation model from the Lake Tahoe Basin.}
\docType{data}
\author{Jonathan A. Greenberg \email{STARStools@estarcion.net}}
\keyword{data}
\examples{tahoe_lidar_bareearth <- raster(system.file("external/tahoe_lidar_bareearth.tif", package="STARStools"))
plot(tahoe_lidar_bareearth)}
