\docType{data}
\name{tahoe_lidar_highesthit.tif}
\alias{tahoe_lidar_highesthit.tif}
\title{Lidar-derived highest hit (aka canopy) digital elevation model from the Lake Tahoe Basin.}
\description{
  Lidar-derived highest hit (aka canopy) digital elevation
  model from the Lake Tahoe Basin.
}
\examples{
tahoe_lidar_highesthit <- raster(system.file("external/tahoe_lidar_highesthit.tif", package="STARStools"))
plot(tahoe_lidar_highesthit)
}
\author{
  Jonathan A. Greenberg \email{STARStools@estarcion.net}
}
\keyword{data}
\keyword{datasets}

