\name{train_test_sampling}
\alias{train_test_sampling}
\title{train_test_sampling}
\usage{
  train_test_sampling(data, factor, train_p = 0.5,
    test_p = 0.5)
}
\arguments{
  \item{data}{TODO}

  \item{factor}{TODO}

  \item{train_p}{TODO}

  \item{test_p}{TODO}
}
\description{
  train_test_sampling
}

