\name{write_raster_list}
\alias{write_raster_list}
\title{write_raster_list}
\usage{
  write_raster_list(raster_list, prefix, outdirectory, ...)
}
\arguments{
  \item{raster_list}{TODO}

  \item{prefix}{TODO}

  \item{outdirectory}{TODO}

  \item{...}{TODO}
}
\description{
  write_raster_list
}

